fix update_camera mode type

+++ samples/model_shader.rs
@@ -49,7 +49,7 @@ fn main() {
     let model_position = Vector3::new(0.0, 0.0, 0.0);
 
     while !rl.window_should_close() {
-        rl.update_camera(&mut camera, CameraMode::CAMERA_FREE as i32);
+        rl.update_camera(&mut camera, CameraMode::CAMERA_FREE);
 
         let mut drawing = rl.begin_drawing(&thread);
         drawing.clear_background(Color::WHITE);
